+++ src/NzbDrone.Core/Jobs/TaskState.cs
@@ -1,0 +1,9 @@
+namespace NzbDrone.Core.Jobs
+{
+    public enum TaskState
+    {
+        Enabled = 0,
+        Disabled = 1,
+        Deleted = 2
+    }
+}