+++ src/Readarr.Api.V1/System/Tasks/TaskStateResource.cs
@@ -1,0 +1,9 @@
+using Readarr.Http.REST;
+
+namespace Readarr.Api.V1.System.Tasks
+{
+    public class TaskStateResource : RestResource
+    {
+        public string State { get; set; }
+    }
+}